[v1.1] high precision output for 'time' and 'times'
This applies an updated version of a patch from @JohnoKing that
increases the default precision of the 'time' and 'times' command
output to six digits after the decimal point.

A new %C formatting specifier is also supported for $TIMEFORMAT --
it is the total of user and system CPU time used.