sh_exec(): do not save/restore PWD for non-BLT_ENV builtins
The question-everything department reporting for duty once again.

Version 2005-05-22 ksh93q+ introduced code that saves the inode and
path of the present working directory before invoking a built-in
command without the BLT_ENV attribute (see data/builtins.c). When
the built-in completes, it gets the PWD's inode again and compares.
If they're different, it uses chdir(2) to change back to the old
working directory.

The corresponding commit in the ksh93-history repo contains no
relevant entries in src/cmd/ksh93/RELEASE so no form of rationale
for this addition is available.

Changing back to a previous directory by path name is unsafe,
because the directory may have been removed and even replaced by a
completely different one by then. This is a common vector for
symlink attacks.

In the 93v- beta, this code was improved to use fstat(2) and
fchdir(2) to guarantee changing back to the same directory.

But is this worth doing at all? Why should a built-in not be able
to change the current working directory? If it's not intended to do
this, it simply should not do it. Even in the case of dynamically
loadable third-party built-ins, we're running built-in code in the
same process as ksh, so we've already decided the code is trusted.
If it's not, there are far worse things than $PWD to worry about.

Not only that, this code comes at a performance hit as the file
system is accessed before and after running a non-BLT_ENV builtin.

Before removing this:

$ arch/*/bin/ksh.old -c 'typeset -i i; \
  time for((i=0;i<=100000;i++)); do sleep 0; done >/dev/null'

real	0m00.863135s
user	0m00.428718s
sys	0m00.433581s

After removing this:

$ arch/*/bin/ksh -c 'typeset -i i; \
  time for((i=0;i<=100000;i++)); do sleep 0; done >/dev/null'

real	0m00.263084s
user	0m00.262078s
sys	0m00.000632s

Removing this nonsense causes no regressions -- which is obvious.
because none of our built-ins except 'cd' change the PWD.

(On a side note, I should probably change the new default
$TIMEFORMAT [re: a1af93f3] to show three digits after the decimal
point. Six is excessive, especially as a default.)