Fix build on macOS (re: 4afc2e1f)
macOS doesn't like it if ast.h is included in vmmopen.c
(conflicting type declarations for brk() and sbrk() between that
file and the OS header), so only include ast.h when we're actually
using the the NoN stub macro from it.

Thanks to Mark Wilson for the build failure report.

Resolves: https://github.com/ksh93/ksh/issues/499