Do not use noreturn in C99 mode on Solaris Sudio cc (re: c4f980eb)
If the -xc99 or -std=c99 flag is passed to Solaris Studio cc, the
build fails with a syntax error on the 'noreturn' keyword. That
keyword was introduced in the C11 standard; C99 does not have it.

The features/common test decides that we can use the 'noreturn'
optimization if the <stdnoreturn.h> header is present on the
system, but that is not correct; standards flags may disable it.

src/lib/libast/features/common:
- Remove two unused tests for extern and void*, all part of C90
  which we now require (re: a34e8319).
- Add test that checks for 'noreturn' by compiling a test program.
- Use that test's result to decide whether to define 'noreturn' as
  empty or not.